backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
motor
passlib[argon2]
python-jose[cryptography]
python-multipart
python-dotenv